# Remove tactic compatibility layer