… dev/doc/changes.txt
… plugins/firstorder/ground.ml
… plugins/firstorder/instances.ml
… plugins/firstorder/rules.ml
… plugins/firstorder/sequent.ml
… plugins/firstorder/sequent.mli
… plugins/firstorder/unify.ml
… plugins/fourier/fourierR.ml
… plugins/ltac/g_rewrite.ml4
… plugins/ltac/rewrite.ml
… plugins/micromega/coq_micromega.ml
… plugins/omega/coq_omega.ml
… plugins/romega/const_omega.ml
… plugins/romega/const_omega.mli
… plugins/romega/g_romega.ml4
… plugins/romega/refl_omega.ml
… proofs/goal.mli
… proofs/refiner.ml
… proofs/refiner.mli
… tactics/auto.ml
… tactics/class_tactics.ml
… tactics/class_tactics.mli
… tactics/equality.ml
… tactics/leminv.ml
… tactics/tacticals.ml
… tactics/tacticals.mli
… tactics/tactics.ml
… vernac/command.ml