This PR removes a lot of compatibility layers in the tactics defined in Coq source code, mainly plugins, that were requiring a cleanup. A few calls to the old engine were also removed, notably in omega.

Overall, the code should be cleaner and potentially more efficient, as the compatibility layer keeps normalizing terms for nothing.

Hopefully, it should not change the semantics of tactics. The `fiat-parsers` travis build is broken, but this is due to an API fix and I have a two-line patch that makes it go through.