@Zimmi48 2017-04-17
@ejgallego's idea with these overlays was that the `dev/ci/ci-user-overlay.sh` file would grow, little by little, to finally contain a history of the compatibility layers that we needed when tweaking Coq. So, if we want to follow this idea, the commit adding the Travis overlay should add a new overlay, not replace the one that was merged from a previous PR.