@ppedrot 2017-04-17
@Zimmi48 Was it designed as such? When merging the EConstr branch, @maximedenes explicitly removed the additional overlay from the merge commit, and I think @ejgallego was overseeing this. Thus I thought the specific overlay was a merge overlook.

Furthermore, I highly doubt that we will get any long-term historical information more than garbage by keeping the overlays in master. Indeed, most of the branches it refers to are going to eventually disappear... If that's not tested, that's dead, bitrotting code.