@Zimmi48 2017-04-17
This is probably best answered by @ejgallego himself (see https://github.com/coq/coq/pull/565#issuecomment-294493854) although my understanding was based on what he explained to me on Gitter (can't find the message, I don't know how to search efficiently for a message in Gitter's history).

I fully agree with you on the garbage problem. That's why I'm suggesting a different approach here https://github.com/coq/coq/pull/565#discussion_r111743935.

But to conclude, I don't really care about what approach we choose, as long as it is kept consistent (if we keep history, try to do it as much as possible, if we don't do it, try not to merge commits adding overlays).