@SkySkimmer 2017-04-17
@Zimmi48 Do you mean https://gitter.im/coq/coq?at=58e798ecad849bcf425f15d2 ?