@ejgallego 2017-04-17
Regarding overlays anything is fine with me, IIRC Maxime doesn't like the garbage so that is why he removes old overlays, I don't think that is a problem as the file is versioned anyways (so when looking for the merge you'll find the info).