@maximedenes 2017-04-19
I remove overlays because I think 1) the garbage will become too huge after a while and 2) the branches it will refer to will no longer exist so it won't be so useful. But I'm ok with changing this process, we can have the discussion somewhere else.