@ppedrot 2017-04-20
Timings:
```
┌────────────────────────┬─────────────────────────┬─────────────────────────────────────┬─────────────────────────────────────┐
│                        │      user time [s]      │             CPU cycles              │          CPU instructions           │
│                        │                         │                                     │                                     │
│           package_name │     NEW     OLD PDIFF   │           NEW           OLD PDIFF   │           NEW           OLD PDIFF   │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼─────────────────────────────────────┤
│              coq-color │  591.23  616.28 -4.06 % │ 1639175953397 1709589416322 -4.12 % │ 2062084426852 2155684583585 -4.34 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼─────────────────────────────────────┤
│              coq-flocq │   53.82   55.22 -2.54 % │  147197283939  150892718589 -2.45 % │  189921753643  193728480646 -1.96 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼─────────────────────────────────────┤
│ coq-mathcomp-odd_order │ 1374.00 1391.83 -1.28 % │ 3827997500511 3875670066025 -1.23 % │ 6788732053877 6824550904241 -0.52 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼─────────────────────────────────────┤
│                 coq-sf │   43.32   43.83 -1.16 % │  118954611173  119462553897 -0.43 % │  154990615632  155419367343 -0.28 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼─────────────────────────────────────┤
│         coq-coquelicot │   70.20   70.70 -0.71 % │  193325481904  194219215927 -0.46 % │  243387052349  244148778915 -0.31 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼─────────────────────────────────────┤
│               coq-hott │  233.36  234.90 -0.66 % │  627718757081  634431876382 -1.06 % │ 1073806098456 1074415791799 -0.06 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼─────────────────────────────────────┤
│  coq-mathcomp-fingroup │   57.19   57.42 -0.40 % │  157358183527  157853795575 -0.31 % │  222310702128  222505853653 -0.09 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼─────────────────────────────────────┤
│  coq-mathcomp-solvable │  187.55  188.28 -0.39 % │  520501470711  522856509381 -0.45 % │  762900838027  765447352977 -0.33 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼─────────────────────────────────────┤
│        coq-fiat-crypto │ 1743.86 1747.52 -0.21 % │ 4841402766474 4857999558016 -0.34 % │ 8089513215483 8121221974639 -0.39 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼─────────────────────────────────────┤
│   coq-mathcomp-algebra │  166.87  167.21 -0.20 % │  463722845250  464378120315 -0.14 % │  664366752289  664513582080 -0.02 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼─────────────────────────────────────┤
│            coq-unimath │ 1243.17 1243.53 -0.03 % │ 3444826114145 3447604610348 -0.08 % │ 5832165556465 5830456345715 +0.03 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼─────────────────────────────────────┤
│ coq-mathcomp-character │  265.06  265.09 -0.01 % │  737182173372  737195299760 -0.00 % │ 1092580280579 1093471293389 -0.08 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼─────────────────────────────────────┤
│     coq-mathcomp-field │  450.20  449.88 +0.07 % │ 1253897741030 1252004654414 +0.15 % │ 2096136833323 2097449067431 -0.06 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼─────────────────────────────────────┤
│                coq-vst │ 1491.00 1489.64 +0.09 % │ 4139255782901 4137784521251 +0.04 % │ 6258921931720 6263916437640 -0.08 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼─────────────────────────────────────┤
│           coq-compcert │  793.21  790.46 +0.35 % │ 2200284214099 2196161056710 +0.19 % │ 3347808519565 3357327116180 -0.28 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼─────────────────────────────────────┤
│               coq-iris │  434.19  432.32 +0.43 % │ 1201290542939 1200636771034 +0.05 % │ 1728232389034 1731478673167 -0.19 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼─────────────────────────────────────┤
│ coq-mathcomp-ssreflect │   42.38   42.04 +0.81 % │  115787015205  115446468563 +0.29 % │  149472230508  149238593626 +0.16 % │
└────────────────────────┴─────────────────────────┴─────────────────────────────────────┴─────────────────────────────────────┘
NEW = b925382975b5218d30aab205cefde6e904a7256f
OLD = 0147ae6ba6db24d4f9b29ff477d374a6abb103dd
```

The bench failed with fiat-parsers, in addition to mathcomp-finmap and geocoq which are already broken in trunk. I don't know if this has been introduced by this branch or if some patch is missing between trunk and this PR.