@ppedrot 2017-04-20
Nevermind, I'm stupid, fiat-parser failed because I did not use the right overlay. I specifically said I had tested it in a message above.