@ppedrot 2017-04-24
New timings:
```
┌────────────────────────┬─────────────────────────┬─────────────────────────────────────┬─────────────────────────────────────┐
│                        │      user time [s]      │             CPU cycles              │          CPU instructions           │
│                        │                         │                                     │                                     │
│           package_name │     NEW     OLD PDIFF   │           NEW           OLD PDIFF   │           NEW           OLD PDIFF   │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼─────────────────────────────────────┤
│              coq-color │  602.23  622.86 -3.31 % │ 1671831569715 1728432952071 -3.27 % │ 2091385753593 2184510917947 -4.26 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼─────────────────────────────────────┤
│              coq-flocq │   55.18   56.25 -1.90 % │  151164126370  153847349537 -1.74 % │  195026816514  198409842467 -1.71 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼─────────────────────────────────────┤
│         coq-coquelicot │   71.04   71.36 -0.45 % │  195968308369  195874482481 +0.05 % │  247479684402  247341588191 +0.06 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼─────────────────────────────────────┤
│ coq-mathcomp-ssreflect │   42.85   43.04 -0.44 % │  116282369973  116658359802 -0.32 % │  150556642415  150378301462 +0.12 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼─────────────────────────────────────┤
│                 coq-sf │   43.70   43.87 -0.39 % │  119989890993  119895800758 +0.08 % │  156120822260  156546972607 -0.27 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼─────────────────────────────────────┤
│           coq-compcert │  815.85  818.82 -0.36 % │ 2263603804452 2273363304342 -0.43 % │ 3391817489027 3403778500640 -0.35 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼─────────────────────────────────────┤
│  coq-mathcomp-fingroup │   57.76   57.95 -0.33 % │  158886637965  158932008256 -0.03 % │  224982351477  225157149003 -0.08 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼─────────────────────────────────────┤
│               coq-iris │  437.69  438.61 -0.21 % │ 1214176969723 1217836354866 -0.30 % │ 1746005783911 1748345322629 -0.13 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼─────────────────────────────────────┤
│  coq-mathcomp-solvable │  191.17  191.05 +0.06 % │  529876063916  530023841181 -0.03 % │  772539109816  774527989759 -0.26 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼─────────────────────────────────────┤
│     coq-mathcomp-field │  453.79  452.74 +0.23 % │ 1262625610273 1259835170213 +0.22 % │ 2107984573611 2109140868689 -0.05 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼─────────────────────────────────────┤
│   coq-mathcomp-algebra │  168.87  168.43 +0.26 % │  468908771865  467883283846 +0.22 % │  669799247250  670910532385 -0.17 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼─────────────────────────────────────┤
│                coq-vst │ 1509.96 1505.59 +0.29 % │ 4195262598022 4181167284995 +0.34 % │ 6302838375418 6306049714401 -0.05 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼─────────────────────────────────────┤
│ coq-mathcomp-odd_order │ 1405.31 1399.66 +0.40 % │ 3916308933937 3893084833749 +0.60 % │ 6834188128762 6871338178275 -0.54 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼─────────────────────────────────────┤
│            coq-unimath │ 1262.25 1256.00 +0.50 % │ 3499533757518 3481636463754 +0.51 % │ 5925558765912 5926520070280 -0.02 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼─────────────────────────────────────┤
│ coq-mathcomp-character │  270.24  268.66 +0.59 % │  751958511956  746779304849 +0.69 % │ 1105211512751 1106451620793 -0.11 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼─────────────────────────────────────┤
│        coq-fiat-crypto │ 1771.18 1759.80 +0.65 % │ 4922360332032 4892713825086 +0.61 % │ 8182455415364 8180524249195 +0.02 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼─────────────────────────────────────┤
│               coq-hott │  235.25  233.26 +0.85 % │  634624358875  629652123730 +0.79 % │ 1077341192623 1077360276042 -0.00 % │
└────────────────────────┴─────────────────────────┴─────────────────────────────────────┴─────────────────────────────────────┘
NEW = b4ca718e7cb62933ea39b61ed3737f3479517bdf
OLD = e57074289193b0f0184f3c7143d8ab7e0edd5112
```

They are similar, thus the stray optimization was probably not observable.